[update:attacher] add attach call.
Description:
support call target process's function directly.

Major Changes:
1. add attach::call.

+++ src/include/libobject/core/utils/string.h
@@ -1,7 +1,7 @@
 #ifndef __CSTRING_H__
 #define __CSTRING_H__
 
-int str_hex_to_int(char *str);
+long str_hex_to_int(char *str);
 char *str_trim(char *str);
 
 #endif
